requirements/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
amqp==5.1.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
async-timeout==4.0.2
asyncio==3.4.3
attrs==21.4.0
billiard==3.6.4.0
boto3==1.21.18
botocore==1.24.18
cached-property==1.5.2
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.0.3
django-debug-toolbar==3.2.4
django-js-asset==2.0.0
django-mptt==0.13.4
django-robots==5.0
django-storages==1.12.3
factory-boy==3.2.1
Faker==13.3.1
frozenlist==1.3.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
isodate==0.6.1
jmespath==0.10.0
kombu==5.2.4
lxml==4.8.0
multidict==6.0.2
packaging==21.3
Pillow==9.0.1
platformdirs==2.5.1
pluggy==1.0.0
prompt-toolkit==3.0.28
psycopg2==2.9.3
py==1.11.0
pycparser==2.21
pyparsing==3.0.7
pytest==7.0.1
pytest-django==4.5.2
pytest-factoryboy==2.1.0
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.19.2
pytz==2021.3
redis==4.1.4
requests==2.27.1
requests-file==1.5.1
requests-toolbelt==0.9.1
s3transfer==0.5.2
sentry-sdk==1.5.7
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.0
yarl==1.7.2
zeep==4.1.0
